Split CLI runtime startup helpers
